freedreno: Fix missing alloc for vertex shader pixsize
The simplest vertex shader was missing ALLOC PARAM/PIXEL SIZE(0x0) which,
on A205 at least, causes GPU hang.

Generate the alloc at the end of compilation pass if not already done.

Clarify the code by dropping the setting of num_position and num_param to zero
and replace with explicit state variables.

Signed-off-by: Martin Fuzzey <mfuzzey@parkeon.com>